[dns] rename to dns_client
